Location Integers changed to Location Bytes
All functionality remains the same, but numbers have been compressed to
bytes.  At the time of writing, SnakeApp works as well.  Now, the size
of the arena is restricted to 127x127 and the number of snakes in an
arena at one time is restricted to 117.

+++ src/SnakeManager.java
@@ -108,8 +108,8 @@ public class SnakeManager {
 	 * Finds the lowest snake ID not currently occupied by a snake
 	 * @return the lowest snake ID not currently occupied by a snake
 	 */
-	public synchronized int getUniqueSnakeID(){
-		int id = -1;
+	public synchronized byte getUniqueSnakeID(){
+		byte id = -1;
 		boolean hasMatch = true;
 		while(hasMatch){
 			id ++;
